requirements.txt
Django==3.0
django-environ==0.4.5
djangorestframework==3.11.0
django-cors-headers==3.7.0
psycopg2-binary>=2.8